Some renaming + clean-up
