+++ src/aac/ui/models_tree.py
@@ -1,4 +1,5 @@
-from tkinter import Canvas, Entry, Tk, LEFT
+from tkinter import Canvas, Entry, Tk, LEFT, GROOVE
+from tkinter.ttk import Frame
 
 
 def add_models_tree(root_window: Tk, view_width: int) -> None:
@@ -9,9 +10,9 @@ def add_models_tree(root_window: Tk, view_width: int) -> None:
         root_window: The window to attach the models tree canvas to
         view_width (int): The width of the canvas
     """
-    models_tree_canvas = Canvas(root_window, width=view_width, height=root_window.winfo_screenheight(), bg='red')
+    models_tree_canvas = Frame(root_window, width=view_width, height=root_window.winfo_screenheight(), borderwidth=1, relief=GROOVE)
     models_tree_canvas.pack(side=LEFT)
 
-    models_tree_entry = Entry(root_window)
-    models_tree_canvas.create_window(view_width // 2, 20, window=models_tree_entry)
+    # models_tree_entry = Entry(root_window)
+    # models_tree_canvas.create_window(view_width // 2, 20, window=models_tree_entry)
